+++ CHANGELOG.md
@@ -4,7 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [unreleased]
 
-- none
+### fixed
+
+- Self-alignment of `Scrollable` elements.
 
 # 0.1.1 (2024-07-18)